Move working sponsored gas code to CastVote
